# api_gateway/docs/swagger.yaml
definitions:
...
  utils.AddProduct:
    properties:
      brandid:
        type: integer
      categoryid:
        type: integer
      image:
        type: string
      modelname:
        type: string
    required:
    - categoryid
    - image
    - modelname
    type: object
  utils.AddProductDetail:
    properties:
      availablecolourid:
        type: integer
      availablesizeid:
        type: integer
      discountid:
        type: integer
      price:
        type: integer
      productid:
        type: integer
      stock:
        type: integer
    required:
    - availablecolourid
    - availablesizeid
    - price
    - productid
    - stock
    type: object
...
  utils.Response:
    properties:
      response:
        type: string
    type: object
host: localhost:4000
info:
  contact: {}
  description: SPORTZONE_E-COMMERCE MICROSERVICE API built using Go, PSQL, REST API
    following Clean Architecture.
  title: E-COMMERCE MICROSERVICE
  version: "2.0"
paths:
...
  /admin/order/detail:
    get:
      consumes:
      - application/json
      description: Users can the selected order details.
      parameters:
      - description: Enter the order id
        in: query
        name: orderid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR VIEWING ORDERS DETAILS
      tags:
      - ADMIN USER
  /admin/product:
    get:
      consumes:
      - application/json
      description: LISTING ALL PRODUCTS FROM ADMINS AND USERS END
      parameters:
      - description: Enter the page number to display
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR LISTING ALL PRODUCTS
      tags:
      - ADMIN USER
  /admin/product/add:
    post:
      consumes:
      - application/json
      description: ADDING PRODUCT FROM ADMINS END
      operationId: ADMIN-ADD-PRODUCT
      parameters:
      - description: Enter the product details
        in: body
        name: product_details
        schema:
          $ref: '#/definitions/utils.AddProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR ADDING PRODUCT
      tags:
      - ADMIN
  /admin/product/detail/{productid}:
    get:
      consumes:
      - application/json
      description: LISTING ALL PRODUCTS DETAILS FROM ADMINS AND USERS END
      parameters:
      - description: Enter the product id that you would like to see the details of
        in: path
        name: productid
        required: true
        type: string
      - description: Enter the page number to display
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR LISTING PRODUCTS DETAILS BY ID
      tags:
      - ADMIN USER
  /admin/product/detail/add:
    post:
      consumes:
      - application/json
      description: ADDING PRODUCT DETAILS FROM ADMINS END
      operationId: ADMIN-ADD-PRODUCT-DETAILS
      parameters:
      - description: Enter the product details
        in: body
        name: product_details
        schema:
          $ref: '#/definitions/utils.AddProductDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR ADDING PRODUCT DETAILS
      tags:
      - ADMIN
...
  /user/checkout/add:
    get:
      consumes:
      - application/json
      description: Users can place a new order with the cart items.
      operationId: USER-PROCEED-ORDER
      parameters:
      - description: Enter the payment id
        in: query
        name: paymentid
        required: true
        type: string
      - description: Enter the address id
        in: query
        name: addressid
        required: true
        type: string
      - description: If you have a coupon,Enter the coupon code
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR PLACING A NEW ORDER
      tags:
      - USER
...
  /user/orders:
    get:
      consumes:
      - application/json
      description: Users can view all orders.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR VIEWING ORDERS
      tags:
      - USER
  /user/orders/detail:
    get:
      consumes:
      - application/json
      description: Users can the selected order details.
      parameters:
      - description: Enter the order id
        in: query
        name: orderid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR VIEWING ORDERS DETAILS
      tags:
      - ADMIN USER
  /user/products:
    get:
      consumes:
      - application/json
      description: LISTING ALL PRODUCTS FROM ADMINS AND USERS END
      parameters:
      - description: Enter the page number to display
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR LISTING ALL PRODUCTS
      tags:
      - ADMIN USER
  /user/products/{productid}:
    get:
      consumes:
      - application/json
      description: LISTING ALL PRODUCTS DETAILS FROM ADMINS AND USERS END
      parameters:
      - description: Enter the product id that you would like to see the details of
        in: path
        name: productid
        required: true
        type: string
      - description: Enter the page number to display
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR LISTING PRODUCTS DETAILS BY ID
      tags:
      - ADMIN USER
...
swagger: "2.0"